requirements.txt
pyyaml>=6.0
aiohttp>=3.9
openai>=1.0
anthropic>=0.30
pydantic>=2.0
fastapi>=0.115
uvicorn>=0.34
